stop trying to be samrt about it for now

+++ openapi_server/test/__init__.py
@@ -8,6 +8,7 @@ from openapi_server.encoder import JSONEncoder
 
 from openapi_server.database import models
 
+SKIP_DB_TESTS = os.getenv("SKIP_DB_TESTS", True)
 
 class BaseTestCase(TestCase):
     def create_app(self):
@@ -15,18 +16,21 @@ class BaseTestCase(TestCase):
         app = connexion.App(__name__, specification_dir="../openapi/")
         app.app.json_encoder = JSONEncoder
         app.add_api("openapi.yaml", pythonic_params=True)
-        app.app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv(
-            "DATABASE_URL", "sqlite:///:memory:"
-        )
-        app.app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-        models.db.init_app(app.app)
+        if not SKIP_DB_TESTS:
+            app.app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv(
+                "DATABASE_URL"
+            )
+            app.app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+            models.db.init_app(app.app)
         return app.app
 
     def setUp(self):
         # check if db url envvar exists and add db to app
         # always seed the db if created
-        models.db.create_all()
+        if not SKIP_DB_TESTS:
+            models.db.create_all()
 
     def tearDown(self):
-        models.db.session.remove()
-        models.db.drop_all()
+        if not SKIP_DB_TESTS:
+            models.db.session.remove()
+            models.db.drop_all()
